Implemented serde support for InetAddr
This implements optional `serde` support, so far only for InetAddr. It'd
be nice to implement it for other things, but this is minimum
requirement for implementing `configure_me`.

Unresolved question: should we defer to the default implementation or
redirect it to `FromStr`, so that the values in config files are nicer?